requirements.txt
fastapi[standard]
sqlalchemy


langchain
langchain-community
langchain-google-genai
langchain_openai
langchain-text-splitters
langchain_qdrant
python-dotenv
pypdf
